pyproject.toml
[project]
name = "sbsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gitpython>=3.1.46",
    "prometheus-client>=0.23.1",
    "python-dotenv>=1.2.1",
    "watchdog>=6.0.0",
]
